Update to 1.3.1
Fixes usenet my list and changes eta to int64

+++ TorBoxNET/Models/TorrentInfoResult.cs
@@ -56,7 +56,7 @@ namespace TorBoxNET
         public int UploadSpeed { get; set; }
 
         [JsonProperty("eta")]
-        public int Eta { get; set; }
+        public Int64 Eta { get; set; }
 
         [JsonProperty("torrent_file")]
         public bool TorrentFile { get; set; }
